Add rotations around a pivot point within an image in minimalist-game-framework::draw::image_ex
Implemented using a modified version of the idea proposed by Megalng[1],
blitting an image onto a larger surface that is then rotated. I was
attempting to make the rotational pivot behavior close to a Pygame
version of SDL2's `SDL_RenderCopyEx`. Currently, this implementation
requires that the pivot point be within the bounds of the provided
image.

[1]: https://stackoverflow.com/a/46252579