Narrow the specification of string->number conversions in the `number` builtin
This changeset adds an actual grammar for string->number conversions
instead of relying solely on Python's `float` builtin. Removes the edge
cases where valid Python numbers (e.g. 123e2) could be parsed by the
`number` builtin even if that number is an invalid Lumpy number literal.

This changeset also allows "Inf" and "inf" as valid string
representations of the `Inf` number builtin and "NaN" and "nan" as valid
string representations of the `NaN` builtin for the purposes of
string->number conversions.