Use UTF-8 as the canonical encoding for string find and slice operations
This is done with the idea that the underlying character representation
of a string in the host language should be decoupled from the
presentation of string indices in Lumpy. UTF-8 was chosen specifically
as UTF-8 has a single endian-independent byte-representation, and because
UTF-8 would likely be the encoding used if a Lumpy interpreter was ever
written in a low-level language such as C or Rust.