Write rune data to standard text streams for dump* and print* calls
Originally, there was an attempt to write raw binary data to the
standard text streams using the streams' underlying buffer object.
Unfortunately, this lead to a bunch of issues such as requiring more
flush calls than necessary, as well as broken pipe errors when the
interpreter was piped into a program such as head(1). There are
workarounds for these problems, but it is somewhat apparent that the
workarounds were going against the grain of the Python language. Not to
mention that these workarounds may break when the interpreter is finally
ported to Windows, where text and binary streams are treated differently
(unlike Unix).