Add the `new` keyword and expression
This changeset adds the `new` keyword and associated expression.

        let myfoo = new foo {
            .bar = some_bar(),
            .baz = some.baz,
        };

        (new vec2 { .x = 3, .y = 4}).normalized();

The `new` keyword, used in an expression of the form:

    new <metamap-expression> <value-expression>

will produce the value from `<value-expression>` with the value of
`<metamap-expression>` as a metamap. The invocation of function
`vec2::init()` using `new` here:

        let vec2 = {
            "init": function(x, y) {
                return new vec2 {
                    "x": x,
                    "y": y,
                };
            },
        }
        let myvec = vec2::init(3, 4);

is roughly equivalent to the following invocation:

        let vec2 = {
            "init": function(x, y) {
                let self = {
                    "x": x,
                    "y": y,
                };
                setmeta(self.&, vec2);
                return self;
            },
        };
        let myvec = vec2::init(3, 4);

with the upside that one fewer map copy operation ends up being used in
practice. This provides a slightly more performant and significantly
more pleasant looking pattern for initialization over the existing
`setmeta(self.&, type)` pattern.

Prior to this commit `sometype::new()` was the existing naming
convention for type-specific initialization functions. With this
changeset turning the `new` function name into a keyword, existing
`sometype::new` functions have been renamed to `sometype::init`.